Enable some new compiler options (#1707)

+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/NameAllocator.kt
@@ -166,7 +166,7 @@ private val RESERVED_KEYWORDS = KEYWORDS + CROSS_PLATFORM_RESERVED_KEYWORDS
 // TODO change to Name?
 internal class NameAllocator
 private constructor(
-  private val allocatedNames: MutableSet<String>,
+  allocatedNames: Set<String>,
   private val tagToName: MutableMap<Any, String>,
   private val mode: Mode,
 ) {
@@ -201,7 +201,7 @@ private constructor(
   ) : this(
     allocatedNames =
       if (preallocateKeywords) {
-        RESERVED_KEYWORDS.toMutableSet()
+        RESERVED_KEYWORDS
       } else {
         mutableSetOf()
       },
@@ -209,20 +209,40 @@ private constructor(
     mode = mode,
   )
 
+  // TODO use explicit backing field in 2.3.0+
+  private val allocatedNames: MutableSet<String> = allocatedNames.toMutableSet()
+
+  fun allocatedNames(): Set<String> = allocatedNames
+
+  fun reserveName(suggestedName: String, tag: Any = Uuid.random().toString()) {
+    @Suppress("RETURN_VALUE_NOT_USED")
+    newNameImpl(suggestedName, tag, generateNewIfExisting = false)
+  }
+
   /**
    * Return a new name using [suggestion] that will not be a Java identifier or clash with other
    * names. The returned value can be queried multiple times by passing `tag` to
    * [NameAllocator.get].
    */
   fun newName(suggestion: String, tag: Any = Uuid.random().toString()): String {
+    return newNameImpl(suggestion, tag, generateNewIfExisting = true)
+  }
+
+  /**
+   * Return a new name using [suggestion] that will not be a Java identifier or clash with other
+   * names. The returned value can be queried multiple times by passing `tag` to
+   * [NameAllocator.get].
+   */
+  private fun newNameImpl(suggestion: String, tag: Any, generateNewIfExisting: Boolean): String {
     val cleanedSuggestion = toSafeIdentifier(suggestion)
     val result = buildString {
       append(cleanedSuggestion)
       var count = 1
       while (!allocatedNames.add(toString())) {
+        if (!generateNewIfExisting) break
         when (mode) {
-          Mode.UNDERSCORE -> append('_')
-          Mode.COUNT -> {
+          UNDERSCORE -> append('_')
+          COUNT -> {
             deleteRange(cleanedSuggestion.length, length)
             append(++count)
           }
@@ -321,4 +341,8 @@ internal fun toSafeIdentifier(suggestion: String) = buildString {
 
 internal fun NameAllocator.newName(suggestion: Name, tag: Any = Uuid.random().toString()): Name {
   return newName(suggestion.asString(), tag).asName()
+}
+
+internal fun NameAllocator.reserveName(suggestion: Name, tag: Any = Uuid.random().toString()) {
+  return reserveName(suggestion.asString(), tag)
 }
